+++ routes/dataModels.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const router = express.Router();
+const { getSeirPredictions } = require('../controllers/sirModel');
+
+router.get("/SEIR", getSeirPredictions);
+
+module.exports = router;
