+++ routes/index.js
@@ -2,4 +2,5 @@ module.exports = {
   healthCheck: require("./healthCheck"),
   google: require("./google"),
   covid19: require("./covid19"),
+  dataModels: require("./dataModels")
 };
